+++ ksqldb-engine/src/main/java/io/confluent/ksql/physical/scalablepush/ProcessingQueue.java
@@ -1,4 +1,19 @@
-package io.confluent.ksql.physical.scalable_push;
+/*
+ * Copyright 2021 Confluent Inc.
+ *
+ * Licensed under the Confluent Community License (the "License"); you may not use
+ * this file except in compliance with the License.  You may obtain a copy of the
+ * License at
+ *
+ * http://www.confluent.io/confluent-community-license
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OF ANY KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations under the License.
+ */
+
+package io.confluent.ksql.physical.scalablepush;
 
 import io.confluent.ksql.execution.streams.materialization.TableRow;
 import java.util.LinkedList;
@@ -11,7 +26,7 @@ public class ProcessingQueue {
   private final int queueSizeLimit;
   private boolean closed = false;
   private boolean droppedRows = false;
-  private Runnable newRowCallback = () -> {};
+  private Runnable newRowCallback = () -> { };
 
   public ProcessingQueue() {
     this(BLOCKING_QUEUE_CAPACITY);
