+++ ksqldb-engine/src/main/java/io/confluent/ksql/physical/scalablepush/locator/PushLocator.java
@@ -1,0 +1,37 @@
+/*
+ * Copyright 2021 Confluent Inc.
+ *
+ * Licensed under the Confluent Community License (the "License"); you may not use
+ * this file except in compliance with the License.  You may obtain a copy of the
+ * License at
+ *
+ * http://www.confluent.io/confluent-community-license
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OF ANY KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations under the License.
+ */
+
+package io.confluent.ksql.physical.scalablepush.locator;
+
+import java.net.URI;
+import java.util.Set;
+
+public interface PushLocator {
+
+  Set<KsqlNode> locate();
+
+  interface KsqlNode {
+
+    /**
+     * @return {@code true} if this is the local node, i.e. the KSQL instance handling the call.
+     */
+    boolean isLocal();
+
+    /**
+     * @return The base URI of the node, including protocol, host and port.
+     */
+    URI location();
+  }
+}
